Feature avis

See merge request agence.digital/training/memoire.franky.nyambi!4

+++ src/main/java/com/api/gesboo/repository/UtilisateurRepository.java
@@ -1,13 +1,13 @@
 package com.api.gesboo.repository;
 
 
-import com.api.gesboo.entite.User.Utilisateur;
+import com.api.gesboo.entite.Utilisateur;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UtilisateurRepository extends CrudRepository<Utilisateur,Integer> {
+public interface UtilisateurRepository extends CrudRepository<Utilisateur, Long> {
     Optional<Utilisateur> findByEmail(String email);
 }
